Adapting DPKG extractor to handle OPKG format
