+++ extractor/filesystem/os/dpkg/testdata/dpkg/invalid
