+++ extractor/filesystem/os/dpkg/testdata/dpkg/status.d/foo
