+++ extractor/filesystem/os/dpkg/testdata/dpkg/trailingnewlines
