+++ extractor/filesystem/os/dpkg/testdata/opkg/nopackage
