+++ extractor/filesystem/os/dpkg/testdata/opkg/trailingnewlines
